+++ src/components/Message/MessageSimple.tsx
@@ -91,7 +91,7 @@ const MessageSimpleWithContext = <
     ? 'str-chat__message str-chat__message--me str-chat__message-simple str-chat__message-simple--me'
     : 'str-chat__message str-chat__message-simple';
 
-  if (message.type === 'message.read' || message.type === 'message.date') {
+  if (message.customType === 'message.date') {
     return null;
   }
 
@@ -206,7 +206,7 @@ const MemoizedMessageSimple = React.memo(
 ) as typeof MessageSimpleWithContext;
 
 /**
- * MessageSimple - UI component that renders a message and receives functionality and logic from the Message/MessageList components.
+ * The default UI component that renders a message and receives functionality and logic from the MessageContext.
  * @example ./MessageSimple.md
  */
 export const MessageSimple = <