+++ src/components/Message/hooks/useReactionHandler.ts
@@ -1,4 +1,5 @@
 import React, { RefObject, useCallback, useEffect, useRef, useState } from 'react';
+import throttle from 'lodash.throttle';
 
 import { useChannelActionContext } from '../../../context/ChannelActionContext';
 import { StreamMessage, useChannelStateContext } from '../../../context/ChannelStateContext';
@@ -6,7 +7,7 @@ import { useChatContext } from '../../../context/ChatContext';
 
 import type { ReactEventHandler } from '../types';
 
-import type { Reaction, ReactionAPIResponse, ReactionResponse } from 'stream-chat';
+import type { Reaction, ReactionResponse } from 'stream-chat';
 
 import type {
   DefaultAttachmentType,
@@ -36,27 +37,37 @@ export const useReactionHandler = <
   const { channel } = useChannelStateContext<At, Ch, Co, Ev, Me, Re, Us>();
   const { client } = useChatContext<At, Ch, Co, Ev, Me, Re, Us>();
 
+  const toggleReaction = throttle(async (id: string, type: string, add: boolean) => {
+    try {
+      if (add) {
+        await channel.sendReaction(id, { type } as Reaction<Re, Us>);
+      } else {
+        await channel.deleteReaction(id, type);
+      }
+    } catch (error) {
+      // revert to the original message if the API call fails
+      if (message) updateMessage(message);
+    }
+  }, 1000);
+
   return async (reactionType: string, event?: React.BaseSyntheticEvent) => {
     if (event?.preventDefault) {
       event.preventDefault();
     }
 
-    if (!updateMessage || !message || !channel || !client) {
-      console.warn(reactionHandlerWarning);
-      return;
+    if (!message) {
+      return console.warn(reactionHandlerWarning);
     }
 
     let userExistingReaction = (null as unknown) as ReactionResponse<Re, Us>;
-
-    const currentUser = client.userID;
 
     if (message.own_reactions) {
       message.own_reactions.forEach((reaction) => {
         // own user should only ever contain the current user id
         // just in case we check to prevent bugs with message updates from breaking reactions
-        if (reaction.user && currentUser === reaction.user.id && reaction.type === reactionType) {
+        if (reaction.user && client.userID === reaction.user.id && reaction.type === reactionType) {
           userExistingReaction = reaction;
-        } else if (reaction.user && currentUser !== reaction.user.id) {
+        } else if (reaction.user && client.userID !== reaction.user.id) {
           console.warn(
             `message.own_reactions contained reactions from a different user, this indicates a bug`,
           );
@@ -64,31 +75,14 @@ export const useReactionHandler = <
       });
     }
 
-    const originalMessage = message;
-    let reactionChangePromise: Promise<ReactionAPIResponse<At, Ch, Co, Me, Re, Us>>;
-
-    // Make the API call in the background
-    // If it fails, revert to the old message...
-    if (message.id) {
+    try {
       if (userExistingReaction) {
-        reactionChangePromise = channel.deleteReaction(message.id, userExistingReaction.type);
+        await toggleReaction(message.id, userExistingReaction.type, false);
       } else {
-        // add the reaction
-        const messageID = message.id;
-
-        const reaction = { type: reactionType } as Reaction<Re, Us>;
-
-        // this.props.channel.state.addReaction(tmpReaction, this.props.message);
-        reactionChangePromise = channel.sendReaction(messageID, reaction);
+        await toggleReaction(message.id, reactionType, true);
       }
-
-      try {
-        // only wait for the API call after the state is updated
-        await reactionChangePromise;
-      } catch (e) {
-        // revert to the original message if the API call fails
-        updateMessage(originalMessage);
-      }
+    } catch (error) {
+      console.log({ error });
     }
   };
 };